added deployments for libp2p test nodes
